test: split preset patterns test to avoid PR conflicts when adding a test in there
